Change gain from short to int on missed files
Former-commit-id: 060f633ccd4349ac5c9392c8e9756b94a5c13b61
Former-commit-id: ee98d1a6d8e44fffc87f0ca922fa80bc64ebc159

+++ NINA/Profile/Interfaces/IPlateSolveSettings.cs
@@ -32,7 +32,7 @@ namespace NINA.Profile {
         BlindSolverEnum BlindSolverType { get; set; }
         string CygwinLocation { get; set; }
         double ExposureTime { get; set; }
-        short Gain { get; set; }
+        int Gain { get; set; }
         short Binning { get; set; }
         FilterInfo Filter { get; set; }
         PlateSolverEnum PlateSolverType { get; set; }
